requirements.txt
praw>=7.0.0
requests>=2.28.0
pandas>=1.5.0
transformers>=4.20.0
torch>=1.12.0
python-dotenv>=1.0.0
